+++ crates/weaver_resolved_schema/src/v2/mod.rs
@@ -5,16 +5,19 @@ use std::collections::{HashMap, HashSet};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use weaver_semconv::{
+    deprecated::Deprecated,
     group::GroupType,
     v2::{
         attribute_group::AttributeGroupVisibilitySpec, signal_id::SignalId, span::SpanName,
         CommonFields,
     },
 };
+use weaver_version::v2::{RegistryChanges, SchemaChanges, SchemaItemChange};
 
 use crate::v2::{
+    attribute::Attribute,
     attribute_group::AttributeGroup,
-    catalog::Catalog,
+    catalog::{AttributeCatalog, Catalog},
     entity::Entity,
     metric::Metric,
     refinements::Refinements,
@@ -46,6 +49,8 @@ pub struct ResolvedTelemetrySchema {
     pub schema_url: String,
     /// The ID of the registry that this schema belongs to.
     pub registry_id: String,
+    /// Catalog of attributes. Note: this will include duplicates for the same key.
+    pub attribute_catalog: Vec<Attribute>,
     /// The registry that this schema belongs to.
     pub registry: Registry,
     /// Refinements for the registry
@@ -57,9 +62,57 @@ impl ResolvedTelemetrySchema {
     /// Statistics about this schema.
     pub fn stats(&self) -> Stats {
         Stats {
-            registry: self.registry.stats(),
+            registry: self.registry.stats(&self.attribute_catalog),
             refinements: self.refinements.stats(),
         }
+    }
+
+    /// Generate a diff between the current schema (must be the most recent one)
+    /// and a baseline schema.
+    #[must_use]
+    pub fn diff(&self, baseline_schema: &ResolvedTelemetrySchema) -> SchemaChanges {
+        // TODO - get manifests
+        SchemaChanges {
+            registry: self.registry_diff(baseline_schema),
+        }
+    }
+
+    #[must_use]
+    fn registry_diff(&self, baseline_schema: &ResolvedTelemetrySchema) -> RegistryChanges {
+        RegistryChanges {
+            attribute_changes: self.registry_attribute_diff(baseline_schema),
+            attribute_group_changes: diff_signals(
+                &self.registry.attribute_groups,
+                &baseline_schema.registry.attribute_groups,
+            ),
+            entity_changes: diff_signals(
+                &self.registry.entities,
+                &baseline_schema.registry.entities,
+            ),
+            event_changes: diff_signals(&self.registry.events, &baseline_schema.registry.events),
+            metric_changes: diff_signals(&self.registry.metrics, &baseline_schema.registry.metrics),
+            span_changes: diff_signals(&self.registry.spans, &baseline_schema.registry.spans),
+        }
+    }
+
+    #[must_use]
+    fn registry_attribute_diff(
+        &self,
+        baseline_schema: &ResolvedTelemetrySchema,
+    ) -> Vec<SchemaItemChange> {
+        let latest_attributes = self.registry_attribute_map();
+        let baseline_attributes = baseline_schema.registry_attribute_map();
+        diff_signals_by_hash(&latest_attributes, &baseline_attributes)
+    }
+
+    /// Get the registry attributes of the resolved telemetry schema in a fast lookup map.
+    fn registry_attribute_map(&self) -> HashMap<&str, &Attribute> {
+        self.registry
+            .attributes
+            .iter()
+            .filter_map(|r| self.attribute_catalog.attribute(r))
+            .map(|a| (a.key.as_str(), a))
+            .collect()
     }
 }
 
@@ -67,12 +120,14 @@ impl ResolvedTelemetrySchema {
 impl TryFrom<crate::ResolvedTelemetrySchema> for ResolvedTelemetrySchema {
     type Error = crate::error::Error;
     fn try_from(value: crate::ResolvedTelemetrySchema) -> Result<Self, Self::Error> {
-        let (registry, refinements) = convert_v1_to_v2(value.catalog, value.registry)?;
+        let (attribute_catalog, registry, refinements) =
+            convert_v1_to_v2(value.catalog, value.registry)?;
         Ok(ResolvedTelemetrySchema {
             // TODO - bump file format?
             file_format: value.file_format,
             schema_url: value.schema_url,
             registry_id: value.registry_id,
+            attribute_catalog,
             registry,
             refinements,
         })
@@ -95,15 +150,15 @@ fn fix_span_group_id(group_id: &str) -> SignalId {
 pub fn convert_v1_to_v2(
     c: crate::catalog::Catalog,
     r: crate::registry::Registry,
-) -> Result<(Registry, Refinements), crate::error::Error> {
+) -> Result<(Vec<Attribute>, Registry, Refinements), crate::error::Error> {
     // When pulling attributes, as we collapse things, we need to filter
     // to just unique.
-    let attributes: HashSet<attribute::Attribute> = c
+    let attributes: HashSet<Attribute> = c
         .attributes
         .iter()
         .cloned()
         .map(|a| {
-            attribute::Attribute {
+            Attribute {
                 key: a.name,
                 r#type: a.r#type,
                 examples: a.examples,
@@ -421,9 +476,31 @@ pub fn convert_v1_to_v2(
         }
     }
 
+    // Now we need to hunt for attribute definitions
+    let mut attributes = Vec::new();
+    for g in r.groups.iter() {
+        for a in g.attributes.iter() {
+            if let Some(attr) = c.attribute(a) {
+                // Attribute definitions do not have lineage.
+                let is_def = g
+                    .lineage
+                    .as_ref()
+                    .and_then(|l| l.attribute(&attr.name))
+                    .is_none();
+                if is_def {
+                    if let Some(v2) = v2_catalog.convert_ref(attr) {
+                        attributes.push(v2);
+                    } else {
+                        // TODO logic error!
+                    }
+                }
+            }
+        }
+    }
+
     let v2_registry = Registry {
         registry_url: r.registry_url,
-        attributes: v2_catalog.into(),
+        attributes,
         spans,
         metrics,
         events,
@@ -435,15 +512,96 @@ pub fn convert_v1_to_v2(
         metrics: metric_refinements,
         events: event_refinements,
     };
-    Ok((v2_registry, v2_refinements))
+    Ok((v2_catalog.into(), v2_registry, v2_refinements))
+}
+
+/// A trait that defines a signal, used for performing "diff"
+pub trait Signal {
+    /// The id of the signal.
+    fn id(&self) -> &str;
+    /// The common fields for the signal.
+    fn common(&self) -> &CommonFields;
+}
+
+/// Diffs signal registries.
+#[must_use]
+fn diff_signals<T: Signal>(latest: &[T], baseline: &[T]) -> Vec<SchemaItemChange> {
+    let baseline_signals: HashMap<&str, &T> = baseline.iter().map(|s| (s.id(), s)).collect();
+    let latest_signals: HashMap<&str, &T> = latest.iter().map(|s| (s.id(), s)).collect();
+    diff_signals_by_hash(&latest_signals, &baseline_signals)
+}
+
+/// Finds the difference between two signal registries using a hash into the signal id.
+fn diff_signals_by_hash<T: Signal>(
+    latest: &HashMap<&str, &T>,
+    baseline: &HashMap<&str, &T>,
+) -> Vec<SchemaItemChange> {
+    let mut changes: Vec<SchemaItemChange> = Vec::new();
+    for (&signal_id, latest_signal) in latest.iter() {
+        let baseline_signal = baseline.get(signal_id);
+        if let Some(baseline_signal) = baseline_signal {
+            if let Some(deprecated) = latest_signal.common().deprecated.as_ref() {
+                // is this a change from the baseline?
+                if let Some(baseline_deprecated) = baseline_signal.common().deprecated.as_ref() {
+                    if deprecated == baseline_deprecated {
+                        continue;
+                    }
+                }
+
+                match deprecated {
+                    Deprecated::Renamed {
+                        renamed_to: rename_to,
+                        note,
+                    } => {
+                        changes.push(SchemaItemChange::Renamed {
+                            old_name: signal_id.to_owned(),
+                            new_name: rename_to.clone(),
+                            note: note.clone(),
+                        });
+                    }
+                    Deprecated::Obsoleted { note } => {
+                        changes.push(SchemaItemChange::Obsoleted {
+                            name: signal_id.to_owned(),
+                            note: note.clone(),
+                        });
+                    }
+                    Deprecated::Unspecified { note } | Deprecated::Uncategorized { note } => {
+                        changes.push(SchemaItemChange::Uncategorized {
+                            name: signal_id.to_owned(),
+                            note: note.clone(),
+                        });
+                    }
+                }
+            }
+        } else {
+            changes.push(SchemaItemChange::Added {
+                name: signal_id.to_owned(),
+            });
+        }
+    }
+    // Any signal in the baseline schema that is not present in the latest schema
+    // is considered removed.
+    // Note: This should never occur if the registry evolution process is followed.
+    // However, detecting this case is useful for identifying a violation of the process.
+    for (signal_name, _) in baseline.iter() {
+        if !latest.contains_key(signal_name) {
+            changes.push(SchemaItemChange::Removed {
+                name: (*signal_name).to_owned(),
+            });
+        }
+    }
+    changes
 }
 
 #[cfg(test)]
 mod tests {
 
+    use crate::v2::attribute::{Attribute as AttributeV2, AttributeRef};
+    use crate::v2::event::Event;
+    use crate::{attribute::Attribute, lineage::GroupLineage, registry::Group};
     use weaver_semconv::{provenance::Provenance, stability::Stability};
 
-    use crate::{attribute::Attribute, lineage::GroupLineage, registry::Group};
+    use crate::lineage::AttributeLineage;
 
     use super::*;
 
@@ -496,6 +654,8 @@ mod tests {
         ]);
         let mut refinement_span_lineage = GroupLineage::new(Provenance::new("tmp", "tmp"));
         refinement_span_lineage.extends("span.my-span");
+        refinement_span_lineage
+            .add_attribute_lineage("test.key".to_owned(), AttributeLineage::new("span.my-span"));
         let v1_registry = crate::registry::Registry {
             registry_url: "my.schema.url".to_owned(),
             groups: vec![
@@ -548,9 +708,11 @@ mod tests {
             ],
         };
 
-        let (v2_registry, v2_refinements) =
+        let (catalog, v2_registry, v2_refinements) =
             convert_v1_to_v2(v1_catalog, v1_registry).expect("Failed to convert v1 to v2");
         // assert only ONE attribute due to sharing.
+        assert_eq!(catalog.len(), 1);
+        // Assert one attribute shows up, due to lineage.
         assert_eq!(v2_registry.attributes.len(), 1);
         // assert attribute fields not shared show up on ref in span.
         assert_eq!(v2_registry.spans.len(), 1);
@@ -620,6 +782,8 @@ mod tests {
         ]);
         let mut refinement_metric_lineage = GroupLineage::new(Provenance::new("tmp", "tmp"));
         refinement_metric_lineage.extends("metric.http");
+        refinement_metric_lineage
+            .add_attribute_lineage("test.key".to_owned(), AttributeLineage::new("metric.http"));
         let v1_registry = crate::registry::Registry {
             registry_url: "my.schema.url".to_owned(),
             groups: vec![
@@ -672,7 +836,7 @@ mod tests {
             ],
         };
 
-        let (v2_registry, v2_refinements) =
+        let (_, v2_registry, v2_refinements) =
             convert_v1_to_v2(v1_catalog, v1_registry).expect("Failed to convert v1 to v2");
         // assert only ONE attribute due to sharing.
         assert_eq!(v2_registry.attributes.len(), 1);
@@ -746,7 +910,7 @@ mod tests {
             }],
         };
 
-        let (v2_registry, _) =
+        let (_, v2_registry, _) =
             convert_v1_to_v2(v1_catalog, v1_registry).expect("Failed to convert v1 to v2");
         assert_eq!(v2_registry.events.len(), 1);
         if let Some(event) = v2_registry.events.first() {
@@ -805,7 +969,7 @@ mod tests {
             }],
         };
 
-        let (v2_registry, _) =
+        let (_, v2_registry, _) =
             convert_v1_to_v2(v1_catalog, v1_registry).expect("Failed to convert v1 to v2");
         assert_eq!(v2_registry.entities.len(), 1);
         if let Some(entity) = v2_registry.entities.first() {
@@ -838,5 +1002,230 @@ mod tests {
         assert_eq!(v2_schema.file_format, "1.0.0");
         assert_eq!(v2_schema.schema_url, "my.schema.url");
         assert_eq!(v2_schema.registry_id, "my-registry");
+    }
+
+    #[test]
+    fn no_diff() {
+        let mut baseline = empty_v2_schema();
+        baseline.attribute_catalog.push(AttributeV2 {
+            key: "test.key".to_owned(),
+            r#type: weaver_semconv::attribute::AttributeType::PrimitiveOrArray(
+                weaver_semconv::attribute::PrimitiveOrArrayTypeSpec::String,
+            ),
+            examples: None,
+            common: CommonFields {
+                brief: "test brief".to_owned(),
+                note: "test note".to_owned(),
+                stability: Stability::Stable,
+                deprecated: None,
+                annotations: Default::default(),
+            },
+        });
+        baseline.registry.attributes.push(AttributeRef(0));
+        let changes = baseline.diff(&baseline);
+        assert!(changes.is_empty());
+    }
+
+    #[test]
+    fn attribute_diff() {
+        let mut baseline = empty_v2_schema();
+        baseline.attribute_catalog.push(AttributeV2 {
+            key: "test.key".to_owned(),
+            r#type: weaver_semconv::attribute::AttributeType::PrimitiveOrArray(
+                weaver_semconv::attribute::PrimitiveOrArrayTypeSpec::String,
+            ),
+            examples: None,
+            common: CommonFields {
+                brief: "test brief".to_owned(),
+                note: "test note".to_owned(),
+                stability: Stability::Stable,
+                deprecated: None,
+                annotations: Default::default(),
+            },
+        });
+        baseline.registry.attributes.push(AttributeRef(0));
+        let mut latest = empty_v2_schema();
+        latest.attribute_catalog.push(AttributeV2 {
+            key: "test.key".to_owned(),
+            r#type: weaver_semconv::attribute::AttributeType::PrimitiveOrArray(
+                weaver_semconv::attribute::PrimitiveOrArrayTypeSpec::String,
+            ),
+            examples: None,
+            common: CommonFields {
+                brief: "test brief".to_owned(),
+                note: "test note".to_owned(),
+                stability: Stability::Stable,
+                deprecated: Some(Deprecated::Renamed {
+                    renamed_to: "test.key.new".to_owned(),
+                    note: "hated it".to_owned(),
+                }),
+                annotations: Default::default(),
+            },
+        });
+        latest.attribute_catalog.push(AttributeV2 {
+            key: "test.key.new".to_owned(),
+            r#type: weaver_semconv::attribute::AttributeType::PrimitiveOrArray(
+                weaver_semconv::attribute::PrimitiveOrArrayTypeSpec::String,
+            ),
+            examples: None,
+            common: CommonFields {
+                brief: "test brief".to_owned(),
+                note: "test note".to_owned(),
+                stability: Stability::Stable,
+                deprecated: None,
+                annotations: Default::default(),
+            },
+        });
+        latest.registry.attributes.push(AttributeRef(0));
+        latest.registry.attributes.push(AttributeRef(1));
+        let diff = latest.diff(&baseline);
+        assert!(!diff.is_empty());
+        for attr_change in diff.registry.attribute_changes.iter() {
+            match attr_change {
+                SchemaItemChange::Renamed {
+                    old_name,
+                    new_name,
+                    note,
+                } => {
+                    assert_eq!(old_name, "test.key");
+                    assert_eq!(new_name, "test.key.new");
+                    assert_eq!(note, "hated it");
+                }
+                SchemaItemChange::Added { name } => {
+                    assert_eq!(name, "test.key.new");
+                }
+                c => panic!("Unexpected change type: {:?}", c),
+            }
+        }
+    }
+
+    #[test]
+    fn v2_detect_metric_removed() {
+        // Test a user changing a metric name but not using deprecated field.
+        let mut baseline = empty_v2_schema();
+        baseline.registry.metrics.push(Metric {
+            name: "http".to_owned().into(),
+            instrument: weaver_semconv::group::InstrumentSpec::UpDownCounter,
+            unit: "s".to_owned(),
+            attributes: vec![],
+            entity_associations: vec![],
+            common: CommonFields::default(),
+        });
+        let mut latest = empty_v2_schema();
+        latest.registry.metrics.push(Metric {
+            name: "http.renamed".to_owned().into(),
+            instrument: weaver_semconv::group::InstrumentSpec::UpDownCounter,
+            unit: "s".to_owned(),
+            attributes: vec![],
+            entity_associations: vec![],
+            common: CommonFields::default(),
+        });
+        let diff = latest.diff(&baseline);
+        assert!(!diff.is_empty());
+        for change in diff.registry.metric_changes.iter() {
+            match change {
+                SchemaItemChange::Added { name } => {
+                    assert_eq!(name, "http.renamed");
+                }
+                SchemaItemChange::Removed { name } => {
+                    assert_eq!(name, "http");
+                }
+                c => panic!("Unexpected change type: {:?}", c),
+            }
+        }
+    }
+
+    #[test]
+    fn v2_detect_entity_uncategorized_deprecation() {
+        // Test a user deprecating an entity with unknown change type.
+        let mut baseline = empty_v2_schema();
+        baseline.registry.entities.push(Entity {
+            common: CommonFields::default(),
+            r#type: "test.entity".to_owned().into(),
+            identity: vec![],
+            description: vec![],
+        });
+        let mut latest = empty_v2_schema();
+        latest.registry.entities.push(Entity {
+            common: CommonFields {
+                deprecated: Some(Deprecated::Uncategorized {
+                    note: "note".to_owned(),
+                }),
+                ..Default::default()
+            },
+            r#type: "test.entity".to_owned().into(),
+            identity: vec![],
+            description: vec![],
+        });
+        let diff = latest.diff(&baseline);
+        assert!(!diff.is_empty());
+        for change in diff.registry.metric_changes.iter() {
+            match change {
+                SchemaItemChange::Uncategorized { name, note } => {
+                    assert_eq!(name, "test.entity");
+                    assert_eq!(note, "note");
+                }
+                c => panic!("Unexpected change type: {:?}", c),
+            }
+        }
+    }
+
+    #[test]
+    fn v2_detect_event_obsoleted() {
+        // Test a user obsoleting an event.
+        let mut baseline = empty_v2_schema();
+        baseline.registry.events.push(Event {
+            common: CommonFields::default(),
+            name: "test.event".to_owned().into(),
+            attributes: vec![],
+            entity_associations: vec![],
+        });
+        let mut latest = empty_v2_schema();
+        latest.registry.events.push(Event {
+            name: "test.event".to_owned().into(),
+            attributes: vec![],
+            entity_associations: vec![],
+            common: CommonFields {
+                deprecated: Some(Deprecated::Obsoleted {
+                    note: "note".to_owned(),
+                }),
+                ..Default::default()
+            },
+        });
+        let diff = latest.diff(&baseline);
+        assert!(!diff.is_empty());
+        for change in diff.registry.metric_changes.iter() {
+            match change {
+                SchemaItemChange::Obsoleted { name, note } => {
+                    assert_eq!(name, "test.event");
+                    assert_eq!(note, "note");
+                }
+                c => panic!("Unexpected change type: {:?}", c),
+            }
+        }
+    }
+
+    // create an empty schema for testing.
+    fn empty_v2_schema() -> ResolvedTelemetrySchema {
+        ResolvedTelemetrySchema {
+            file_format: "1.0.0".to_owned(),
+            schema_url: "my.schema.url".to_owned(),
+            registry_id: "main".to_owned(),
+            attribute_catalog: vec![],
+            registry: Registry {
+                attributes: vec![],
+                attribute_groups: vec![],
+                registry_url: "todo".to_owned(),
+                spans: vec![],
+                metrics: vec![],
+                events: vec![],
+                entities: vec![],
+            },
+            refinements: Refinements {
+                spans: vec![],
+                metrics: vec![],
+                events: vec![],
+            },
+        }
     }
 }